implement gif output
