+++ src/pages/index.tsx
@@ -105,6 +105,7 @@ const Index = () => {
             )}
             <Script async src="https://platform.twitter.com/widgets.js" />
           </SnsParagraph>
+          <Script src="/gif.js" />
         </Footer>
       </Main>
     </>
